backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
openai>=1.0.0,<2.0.0 # For OpenAI API calls
